requirements.txt
Django>=3.0
psycopg2-binary>=2.8
djangorestframework>=3.13.1
djoser>=2.1.0
markdown
django-filter
drf-yasg